% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp2Data.R
\name{comp2Data}
\alias{comp2Data}
\title{Simulates data for Compulsory 2}
\usage{
comp2Data(nMark = 5, nCh = 2, plotte = F, write = F, seed = NULL)
}
\arguments{
\item{nMark}{Integer. No of markers}

\item{nCh}{Integer. No of children, at least 2}

\item{plotte}{Logical. Plots if TRUE}

\item{write}{Logical. Writes to file if TRUE}

\item{seed}{Integer. Required}
}
\value{
Simulated half sib ped.
}
\description{
Simulates data for half sibs
}
\examples{
library(pedsuite)
comp2Data(seed = 1729)

}
