% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAlleleFrequencies.R
\name{estimateAlleleFrequencies}
\alias{estimateAlleleFrequencies}
\title{Estimates allele frequencies}
\usage{
estimateAlleleFrequencies(x, markers = NULL)
}
\arguments{
\item{x}{a ped object or a list of such}

\item{markers}{If `NULL`, results are given for all markers}
}
\value{
A list, eachelement containing allele frequencies
}
\description{
A wrapper for genetics:::summary.genotype.
}
\examples{
db = simulateUnrelated(nInd = 100, nMark = 2, seed = 1729)
estimateAlleleFrequencies(db)
estimateAlleleFrequencies(db, markers = 1)
}
