% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfSibPedSim.R
\name{halfSibPedSim}
\alias{halfSibPedSim}
\title{Simulates data for male fathering children with multiple females}
\usage{
halfSibPedSim(nMark = 5, nCh = 2, plotte = F, write = F, seed = NULL)
}
\arguments{
\item{nMark}{Integer. No of markers.}

\item{nCh}{Integer. No of children.}

\item{plotte}{Logical. Plots if TRUE.}

\item{write}{Logical. Writes to file if TRUE.}

\item{seed}{Integer. Required.}
}
\value{
Simulated pedigre
}
\description{
Simulates SNP data for father and children
}
\examples{
halfSibPedSim(seed = 1729)

}
