% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleSNP.R
\name{mleSNP}
\alias{mleSNP}
\title{MLE estimate for SNP marker}
\usage{
mleSNP(x, mark = 1, alleles = 1:2)
}
\arguments{
\item{x}{A ped object.}

\item{mark}{Index or name of marker}
}
\value{
Estimate of frequency of first allele.
}
\description{
Estimates allele frequency of first allele in specified marker by maximising the likelihood.
}
\examples{
x = halfSibPedSim(seed = 1729, nCh = 100, nMark = 2)
mleSNP(x, mark = 1)
mleSNP(x, mark = NULL)

}
