% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateUnrelated.R
\name{simulateUnrelated}
\alias{simulateUnrelated}
\title{Simulates marker data for unrelated individuals}
\usage{
simulateUnrelated(
  nMark = 5,
  nInd = 2,
  alleles = NULL,
  afreq = NULL,
  seed = NULL
)
}
\arguments{
\item{nMark}{Integer. No of markers}

\item{nInd}{Integer. No of individuals}

\item{alleles}{designations}

\item{afreq}{frequencies}

\item{seed}{Integer.}
}
\value{
Simulated marker data
}
\description{
A wrapper for `markerSim`.
}
\examples{
simulateUnrelated(seed = 1729)

}
